[QtCreator] Removed hard-coded paths

+++ sw/tools/qtcreator_ide_config/configure_workspace.py
@@ -4,15 +4,24 @@ import sys
 import os
 import time
 
-# Set home directory
-PAPARAZZI_HOME = os.path.realpath(os.path.dirname(os.path.dirname(os.path.realpath(__file__))) + '../../../')
+# Set home directory and unittest directory
+TOOL_DIR       = os.path.dirname(os.path.realpath(__file__))
+PAPARAZZI_HOME = os.path.realpath(os.path.join(TOOL_DIR, os.pardir, os.pardir, os.pardir))
+TEST_HOME      = os.path.join(PAPARAZZI_HOME, "tests", "unittest")
 
 # Aircraft name is argument and should be equal to project name
 AIRCRAFT = sys.argv[1]
 
 # Check if the IDE files exist with that name
-if not os.path.isfile(PAPARAZZI_HOME + "/" + AIRCRAFT + ".creator"):
+if not os.path.isfile(os.path.join(PAPARAZZI_HOME, AIRCRAFT + ".creator")):
 	raise Exception("Aircraft name does not match project name.")
+
+# Check if also a test environment is configured for this aircraft
+if not os.path.isfile(os.path.join(TEST_HOME, AIRCRAFT + "Tester.creator")):
+	print "No test environment configured for this aircraft. To do so, create a project called \"" + AIRCRAFT + "Tester\" in " + TEST_HOME + "."
+	TEST_CONF = False
+else:
+	TEST_CONF = True
 
 # First, do make clean on the aircraft
 cmd = "make -C " + PAPARAZZI_HOME + " -f Makefile.ac AIRCRAFT=" + AIRCRAFT + " clean_ac"
@@ -23,20 +32,21 @@ reply = os.popen(cmd).read()
 print "========================"
 
 # Remove project files
-# If the interpeter runs for the first time, it regenerates these files. Only the first time
-config_file   = PAPARAZZI_HOME + "/" + AIRCRAFT + ".config"
-includes_file = PAPARAZZI_HOME + "/" + AIRCRAFT + ".includes"
-files_file    = PAPARAZZI_HOME + "/" + AIRCRAFT + ".files"
-test_config_file   = PAPARAZZI_HOME + "/tests/unittest/" + AIRCRAFT + "Tester.config"
-test_includes_file = PAPARAZZI_HOME + "/tests/unittest/" + AIRCRAFT + "Tester.includes"
+# If the gcc_interpeter.py runs for the first time, it regenerates these files. Only the first time
+config_file   = os.path.join(PAPARAZZI_HOME, AIRCRAFT + ".config")
+includes_file = os.path.join(PAPARAZZI_HOME, AIRCRAFT + ".includes")
+files_file    = os.path.join(PAPARAZZI_HOME, AIRCRAFT + ".files")
 if os.path.isfile(config_file):
 	os.remove(config_file)
 if os.path.isfile(includes_file):
 	os.remove(includes_file)
-if os.path.isfile(test_config_file):
-	os.remove(test_config_file)
-if os.path.isfile(test_includes_file):
-	os.remove(test_includes_file)
+if TEST_CONF:
+	test_config_file   = os.path.join(TEST_HOME, AIRCRAFT + "Tester.config")
+	test_includes_file = os.path.join(TEST_HOME, AIRCRAFT + "Tester.includes")
+	if os.path.isfile(test_config_file):
+		os.remove(test_config_file)
+	if os.path.isfile(test_includes_file):
+		os.remove(test_includes_file)
 
 # create empty files file
 if os.path.isfile(files_file):
@@ -46,35 +56,46 @@ new_files_file.close()
 
 # Then, do a full build on the aircraft, but replace compiler with python script
 print "Analyzing make output.."
-cmd = "make -f Makefile.ac PAPARAZZI_HOME=" + PAPARAZZI_HOME + " PAPARAZZI_SRC=" + PAPARAZZI_HOME + " AIRCRAFT=" + AIRCRAFT + " CC='python " + PAPARAZZI_HOME + "/sw/tools/qtcreator_ide_config/gcc_interpeter.py " + AIRCRAFT + "' CP='python " + PAPARAZZI_HOME + "/sw/tools/qtcreator_ide_config/skip_command.py' LD='python " + PAPARAZZI_HOME + "/sw/tools/qtcreator_ide_config/skip_command.py' DMP='python " + PAPARAZZI_HOME + "/sw/tools/qtcreator_ide_config/skip_command.py' NM='python " + PAPARAZZI_HOME + "/sw/tools/qtcreator_ide_config/skip_command.py' NPROCS=1 ap.compile"
-print os.popen(cmd).read()
+cmd = "make -f Makefile.ac PAPARAZZI_HOME=" + PAPARAZZI_HOME + " PAPARAZZI_SRC=" + PAPARAZZI_HOME + " AIRCRAFT=" + AIRCRAFT + " CC='python " + os.path.join(TOOL_DIR, "gcc_interpeter.py") + " " + AIRCRAFT + "' CP='python " + os.path.join(TOOL_DIR, "skip_command.py") + "' LD='python " + os.path.join(TOOL_DIR, "skip_command.py") + "' DMP='python " + os.path.join(TOOL_DIR, "skip_command.py") + "' NM='python " + os.path.join(TOOL_DIR, "skip_command.py") + "' NPROCS=1 ap.compile"
+output = os.popen(cmd).read()
+print "Done analyzing make output."
 
-# Get all the defines from config file
-UNITTEST_HOME = PAPARAZZI_HOME + "/tests/unittest"
-test_config_file = open(PAPARAZZI_HOME + "/tests/unittest/" + AIRCRAFT + "Tester.config")
-macro_text = test_config_file.read()
-test_config_file.close()
 
-macro_flags = macro_text.split('\n')
-for i, s in enumerate(macro_flags):
-	macro_flags[i] = s.replace("#define ", "    - -D").replace(" ", "='").replace("='='='='-='-D", "    - -D") + "'" #beun
+## The next section generates the yml file used for unittesting.
+if TEST_CONF:
+	print "========================"
+	print "Configuring unittest framework.."
+	# Get all the defines from config file.
 
-macro_flags.pop()
+	test_config_file = open(os.path.join(TEST_HOME, AIRCRAFT + "Tester.config"))
+	macro_text = test_config_file.read()
+	test_config_file.close()
+	macro_flags = macro_text.split("\n")
+	for i, s in enumerate(macro_flags):
+		# defines in <AIRCRAFT>.yml
+		macro_flags[i] = "    - -D" + s.replace("#define ", "").replace(" ", "='") + "'"
 
-# Get include directories from testing includes file
-test_includes_file = open(UNITTEST_HOME + "/" + AIRCRAFT + "Tester.includes")
-test_includes_content = test_includes_file.read().split('\n')
-for i, s in enumerate(test_includes_content):
-	test_includes_content[i] = "      - '" + s + "/'"
-yml_include_items = '\n'.join(test_includes_content)
+	macro_flags.pop()
+	yml_macro_flags = "\n".join(macro_flags)
 
-# Configure the yml script for unit testing
-original_yml = open(UNITTEST_HOME + "/ymlmodel")
-original_content = original_yml.read()
-original_yml.close()
+	# Get include directories from testing includes file
+	test_includes_file = open(os.path.join(TEST_HOME, AIRCRAFT + "Tester.includes"))
+	test_includes_content = test_includes_file.read().split("\n")
+	test_includes_file.close()
+	for i, s in enumerate(test_includes_content):
+		test_includes_content[i] = "      - '" + s + "/'"
+	
+	test_includes_content.pop()
+	yml_include_items = "\n".join(test_includes_content)
 
-yml_macro_flags = '\n'.join(macro_flags)
-new_content = original_content.replace("MACRO_ITEMS", yml_macro_flags).replace("INCLUDE_ITEMS", yml_include_items)
-new_yml = open(UNITTEST_HOME + "/" + AIRCRAFT + ".yml", "w")
-new_yml.write(new_content)
-new_yml.close()
+	# Configure the yml script for unit testing
+	original_yml = open(os.path.join(TEST_HOME, "template.yml"))
+	original_content = original_yml.read()
+	original_yml.close()
+
+	new_content = original_content.replace("MACRO_ITEMS", yml_macro_flags).replace("INCLUDE_ITEMS", yml_include_items)
+	new_yml = open(os.path.join(TEST_HOME, AIRCRAFT + ".yml"), "w")
+	new_yml.write(new_content)
+	new_yml.close()
+	print "Done completing unittest framework."
+	print "========================"
